+++ radio/ETHERNET/benetel/4g/benetel.c
@@ -243,7 +243,7 @@ next:
   //printf("BENETEL: %s (f.sf %d.%d)\n", __FUNCTION__, *frame, *subframe);
 
   RU_proc_t *proc = &ru->proc;
-  extern uint16_t sf_ahead;
+  extern int sf_ahead;
   int f = *frame;
   int sf = *subframe;
